WWCS/requirements_python.txt
topo-descriptors
numpy
elevation
rasterio
rioxarray
ecmwf-api-client
cdo
requests
pandas
mysql-connector-python
pyyaml
python-dotenv
ecmwf-opendata
